Se hizo todo la capa modelo

+++ Codigo/AkuavidaJPA/src/main/java/edu/co/sena/akuavida/modelo/entitis/Inventario.java
@@ -78,7 +78,7 @@ public class Inventario implements Serializable {
         return cantidadProductoComprado;
     }
 
-    public void setCantidadProductoComprado(float cantidadProductoComprado) {
+    public void setCantidadProductoComprado(int cantidadProductoComprado) {
         this.cantidadProductoComprado = cantidadProductoComprado;
     }
